+++ src/main/java/com/example/cardmarket/repository/ArticleRepository.java
@@ -3,5 +3,5 @@ package com.example.cardmarket.repository;
 import com.example.cardmarket.entity.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<Article, Integer> {
+public interface ArticleRepository extends JpaRepository<Article, Long> {
 }
